A Proper HTML Server: The Commit
Added a 404 Page
Implemented a 404 Handler to deal with the pesky .startsWith context urls
Revised the ServeFile method so that the user has to handle and serve the exact file. Still uses "html/" as it is the root directory, but all other relative pathing is the responsibility of the method caller.
Fixed the bug where you could not access subdirectories without ending them with a "/"